requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncclick==8.1.7.2
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.8
frozenlist==1.4.1
idna==3.7
kasa-crypt==0.4.2
multidict==6.0.5
orjson==3.10.3
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
python-kasa==0.6.2.1
sniffio==1.3.1
typing_extensions==4.12.1
yarl==1.9.4
